Add [FA] adding cancel order and fix all error

+++ database/migrations/2024_04_29_012752_create_menus_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('category');
             $table->string('second_category');
             $table->integer('stock');
-            $table->float('rating')->nullable();
+            $table->double('rating')->nullable();
             $table->text('description');
             $table->boolean('status_menu')->default(true);
             $table->timestamps();
