+++ database/migrations/2024_07_18_002550_create_order_details_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('menu_id');
             $table->unsignedBigInteger('order_id')->nullable();
             $table->unsignedBigInteger('variant_id')->nullable();
-            $table->float('rating')->nullable();
+            $table->double('rating')->nullable();
             $table->timestamps();
 
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('set null');
